Changed the data type mapping for FdoDecimal from Integer to Real. Most of the time Integer is ok, but I got a sample file where the FdoDecimals had stuff after the decimal point... so I had to switch it. This is unfortunate, because it will propagate problems with files where the Decimal is really Integer, but I guess we can enhance the SQLiteConverter to optionally map Decimal to Int type during conversion.
git-svn-id: http://svn.osgeo.org/fdo/trunk@4364 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/SQLite/Provider/SltProvider.cpp
@@ -71,7 +71,7 @@ SltConnection::SltConnection() : m_refCount(1)
         g_fdo2sql_map[FdoDataType_Boolean] =    "INTEGER";
         g_fdo2sql_map[FdoDataType_Byte] =       "INTEGER";
         g_fdo2sql_map[FdoDataType_DateTime] =   "TEXT";
-        g_fdo2sql_map[FdoDataType_Decimal] =    "INTEGER";
+        g_fdo2sql_map[FdoDataType_Decimal] =    "REAL";
         g_fdo2sql_map[FdoDataType_Double] =     "REAL";
         g_fdo2sql_map[FdoDataType_Int16] =      "INTEGER";
         g_fdo2sql_map[FdoDataType_Int32] =      "INTEGER";
